Quiz/requirements.txt

bcrypt==4.3.0
blinker==1.9.0
click
colorama==0.4.6
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
celery==5.3.6
redis==5.0.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
